BC-588: Extract claim details code into package
