+++ src/main/resources/messages.properties
@@ -99,8 +99,8 @@ assessmentOutcome.cancel=Cancel
 provider.firmName.notAvailable=Currently not available
 
 claimDetails.mainHeading=Claim details
-claimDetails.overview=Overview
-claimDetails.claimHistory=Claim history
+claimDetails.overviewHeading=Overview
+claimDetails.claimHistoryHeading=Claim history
 
 claimSummary.summary=Summary
 claimSummary.totalHeading=Totals
