+++ src/main/resources/templates/claimdetails/claim-summary.html
@@ -13,7 +13,7 @@
                 <div class="govuk-grid-column-full">
                     <th:block th:replace="~{partials/claim-details-header :: claimDetailsHeader(currentPage='overview')}"></th:block>
 
-                    <h2 class="govuk-heading-m" th:text="#{claimDetails.overview}"></h2>
+                    <h2 class="govuk-heading-m" th:text="#{claimDetails.overviewHeading}"></h2>
 
                     <th:block th:if="${claim != null}">
                         <div class="govuk-summary-card">
