+++ src/test/java/uk/gov/justice/laa/amend/claim/controllers/ClaimSummaryControllerTest.java
@@ -1,5 +1,7 @@
 package uk.gov.justice.laa.amend.claim.controllers;
 
+import static org.hamcrest.Matchers.hasProperty;
+import static org.hamcrest.Matchers.is;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
 import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.csrf;
@@ -24,6 +26,7 @@ import org.junit.jupiter.params.provider.EnumSource;
 import org.springframework.boot.webmvc.test.autoconfigure.WebMvcTest;
 import org.springframework.mock.web.MockHttpSession;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
+import uk.gov.justice.laa.amend.claim.controllers.claimdetails.ClaimSummaryController;
 import uk.gov.justice.laa.amend.claim.models.AssessmentInfo;
 import uk.gov.justice.laa.amend.claim.models.CivilClaimDetails;
 import uk.gov.justice.laa.amend.claim.models.Role;
@@ -70,7 +73,7 @@ public class ClaimSummaryControllerTest extends BaseControllerTest {
     mockMvc
         .perform(get(buildPath()).session(session))
         .andExpect(status().isOk())
-        .andExpect(view().name("claim-summary"))
+        .andExpect(view().name("claimdetails/claim-summary"))
         .andExpect(model().attributeExists("claim"))
         .andExpect(model().attribute("searchUrl", "/"))
         .andExpect(request().sessionAttribute(claimId.toString(), claim));
@@ -95,7 +98,7 @@ public class ClaimSummaryControllerTest extends BaseControllerTest {
     mockMvc
         .perform(get(buildPath()).session(session))
         .andExpect(status().isOk())
-        .andExpect(view().name("claim-summary"))
+        .andExpect(view().name("claimdetails/claim-summary"))
         .andExpect(model().attributeExists("claim"))
         .andExpect(model().attribute("searchUrl", "/?officeCode=12345&page=1"))
         .andExpect(request().sessionAttribute(claimId.toString(), claim));
@@ -162,7 +165,10 @@ public class ClaimSummaryControllerTest extends BaseControllerTest {
     mockMvc
         .perform(get(buildPath()).session(session))
         .andExpect(status().isOk())
-        .andExpect(model().attribute("isAssessmentButtonPresent", true));
+        .andExpect(
+            model()
+                .attribute(
+                    "claimDetailsFooterView", hasProperty("assessmentButtonPresent", is(true))));
   }
 
   @Test
@@ -181,7 +187,10 @@ public class ClaimSummaryControllerTest extends BaseControllerTest {
     mockMvc
         .perform(get(buildPath()).session(session))
         .andExpect(status().isOk())
-        .andExpect(model().attribute("isAssessmentButtonPresent", false));
+        .andExpect(
+            model()
+                .attribute(
+                    "claimDetailsFooterView", hasProperty("assessmentButtonPresent", is(false))));
   }
 
   @Test
@@ -197,7 +206,10 @@ public class ClaimSummaryControllerTest extends BaseControllerTest {
     mockMvc
         .perform(get(buildPath()).session(session))
         .andExpect(status().isOk())
-        .andExpect(model().attribute("isAssessmentButtonPresent", false));
+        .andExpect(
+            model()
+                .attribute(
+                    "claimDetailsFooterView", hasProperty("assessmentButtonPresent", is(false))));
   }
 
   @Test
@@ -212,7 +224,10 @@ public class ClaimSummaryControllerTest extends BaseControllerTest {
     mockMvc
         .perform(get(buildPath()).session(session))
         .andExpect(status().isOk())
-        .andExpect(model().attribute("isAssessmentButtonPresent", false));
+        .andExpect(
+            model()
+                .attribute(
+                    "claimDetailsFooterView", hasProperty("assessmentButtonPresent", is(false))));
   }
 
   @Test
@@ -228,7 +243,10 @@ public class ClaimSummaryControllerTest extends BaseControllerTest {
     mockMvc
         .perform(get(buildPath()).session(session))
         .andExpect(status().isOk())
-        .andExpect(model().attribute("isAssessmentButtonPresent", true));
+        .andExpect(
+            model()
+                .attribute(
+                    "claimDetailsFooterView", hasProperty("assessmentButtonPresent", is(true))));
   }
 
   @Test
@@ -243,7 +261,9 @@ public class ClaimSummaryControllerTest extends BaseControllerTest {
     mockMvc
         .perform(get(buildPath()).session(session))
         .andExpect(status().isOk())
-        .andExpect(model().attribute("isVoidButtonPresent", true));
+        .andExpect(
+            model()
+                .attribute("claimDetailsFooterView", hasProperty("voidButtonPresent", is(true))));
   }
 
   @Test
@@ -262,7 +282,9 @@ public class ClaimSummaryControllerTest extends BaseControllerTest {
     mockMvc
         .perform(get(buildPath()).session(session))
         .andExpect(status().isOk())
-        .andExpect(model().attribute("isVoidButtonPresent", false));
+        .andExpect(
+            model()
+                .attribute("claimDetailsFooterView", hasProperty("voidButtonPresent", is(false))));
   }
 
   @Test
@@ -281,7 +303,9 @@ public class ClaimSummaryControllerTest extends BaseControllerTest {
     mockMvc
         .perform(get(buildPath()).session(session))
         .andExpect(status().isOk())
-        .andExpect(model().attribute("isVoidButtonPresent", false));
+        .andExpect(
+            model()
+                .attribute("claimDetailsFooterView", hasProperty("voidButtonPresent", is(false))));
   }
 
   private String buildPath() {
